requirements.txt
dotevn==0.0.5
discord.py==1.6.*
